Use init.js to synchronously add deferred scripts
This is a different approach to the previous commit, which did not work
as it resulted in scripts being loaded in a random order.

This way, we synchronously run init.js (with its own random seed) which
adds the other scripts in the correct order as deferred-load script
references, which can then be loaded in the normal way by the browser -
but also with a seed so we always get the latest version of the script.